The Install and License functions now work.
The Install and License functions will now correctly install and license
the PrizmDoc server and client. I've only tested these on Win10 so far,
so more testing should prooooooobably be done.

The Remove function does not remove the programs from the system, it
just deletes the files from the computer, so this should probably be
updated in the future.